Handle OTP failure flow and transaction status updates with Kafka integration
- Mark transaction as FAILED when OTP attempts exceed limit
- Add custom exceptions for OTP and transaction failures
- Publish transaction failure events to Kafka after commit
- Add Kafka topic configuration and consumer/producer setup
- Improve OTP verification and attempt tracking logic
- Update repositories with locking queries
- Fix and stabilize controller, service, and repository tests

+++ customer-microservice/src/main/java/com/ritik/customer_microservice/model/OtpVerification.java
@@ -34,7 +34,7 @@ public class OtpVerification {
     private LocalDateTime expiryTime;
 
     @Column(name="attempt_count",nullable = false)
-    private byte attemptCount;
+    private int attemptCount;
 
     @Column(name = "verified", nullable = false)
     private boolean verified;
